Add validation to exclude direct subcase migration
Fixes #393
